Add files via upload
Python, API (for fetching a random word), API(Oxford API) for getting the definition, Desktop Notification(plyer) the notification stays for a specific amount of time and goes away we can give the time and if run it in the background for every hour it pops up and stays for few seconds and goes away so you are learning new every hour. It is very useful for GRE aspirants as they need to learn a lot of new words every day.